Moved apex file parsing to its own class
